Extract common modules to subdirectory
