+++ client/src/AppLayout.js
@@ -1,10 +1,13 @@
 import { useContext } from 'react';
+import { Redirect, Route } from "react-router-dom";
 
+import AnimatedRouter from './lib/AnimatedRouter';
+import AuthenticatedRoute from './lib/AuthenticatedRoute';
+import BrandBar from './lib/BrandBar';
+import Footer from './lib/Footer';
+import NavItems from './NavItems';
 import styles from './AppLayout.module.css';
-import AnimatedRouter from './AnimatedRouter';
-import BrandBar from './BrandBar';
-import Footer from './Footer';
-import { Context as ConfigContext } from './ConfigContext';
+import { Context as ConfigContext } from './lib/ConfigContext';
 import { routes, unconfiguredRoutes } from './routes';
 
 function AppLayout() {
@@ -12,7 +15,7 @@ function AppLayout() {
 
   return (
     <>
-    <BrandBar />
+    <BrandBar navItems={<NavItems />} />
     <div
       className="container-fluid"
       id="main"
@@ -20,6 +23,9 @@ function AppLayout() {
       <div id="toast-portal" className={styles.ToastPortal}></div>
       <div className="content">
         <AnimatedRouter
+          AuthenticatedRoute={AuthenticatedRoute}
+          Redirect={Redirect}
+          Route={Route}
           exact={!unconfigured}
           routes={unconfigured ? unconfiguredRoutes : routes}
           sideNav={SideNav}
