+++ client/src/NavItems.js
@@ -1,34 +1,7 @@
-import React, { useContext } from 'react';
+import { useContext } from 'react';
 import { Link } from "react-router-dom";
-import classNames from 'classnames';
 
-import AccountMenu from './account/Menu';
-import { BrandbarLogo, BrandbarText } from './Branding';
 import { Context as CurrentUserContext } from './account/CurrentUserContext';
-
-export default function BrandBar({ className }) {
-  return (
-    <nav className={classNames('navbar navbar-expand-lg navbar-light bg-white', className)}>
-      <a
-        className="navbar-brand"
-        href="/"
-      >
-        <BrandbarLogo />
-      </a>
-      <BrandbarText />
-
-      <div className="collapse navbar-collapse">
-        <ul className="navbar-nav mr-auto">
-          <NavItems />
-        </ul>
-        <ul className="navbar-nav">
-          <AccountMenu />
-        </ul>
-      </div>
-
-    </nav>
-  );
-}
 
 function NavItems() {
   const { currentUser } = useContext(CurrentUserContext);
@@ -63,3 +36,5 @@ function NavItems() {
     </>
   );
 }
+
+export default NavItems;
