+++ client/src/lib/Footer.js
