+++ client/src/lib/UnauthorizedError.js
