+++ client/src/lib/png_trans_logo-navbar.png
